+++ api/v1alpha1/constant.go
@@ -36,3 +36,27 @@ const (
 	// DefaultMode is exclusive
 	DefaultMode PlacementMode = ""
 )
+
+type NetworkMode string
+
+const (
+	HostNetworkMode NetworkMode = "HostNetwork"
+
+	ContainerNetworkMode NetworkMode = "ContainerNetwork"
+
+	// DefaultMode is Host
+	DefaultNetworkMode NetworkMode = ""
+)
+
+type FuseCleanPolicy string
+
+const (
+	// NoneCleanPolicy is the default clean policy. It will be transformed to OnRuntimeDeletedCleanPolicy automatically.
+	NoneCleanPolicy FuseCleanPolicy = ""
+
+	// OnDemandCleanPolicy cleans fuse pod once th fuse pod on some node is not needed
+	OnDemandCleanPolicy FuseCleanPolicy = "OnDemand"
+
+	// OnRuntimeDeletedCleanPolicy cleans fuse pod only when the cache runtime is deleted
+	OnRuntimeDeletedCleanPolicy FuseCleanPolicy = "OnRuntimeDeleted"
+)