+++ pkg/ddc/alluxio/operations/conf_test.go
@@ -1,11 +1,29 @@
+/*
+Copyright 2021 The Fluid Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package operations
 
 import (
 	"errors"
-	"github.com/brahma-adshonor/gohook"
-	"github.com/fluid-cloudnative/fluid/pkg/utils/kubeclient"
 	"strings"
 	"testing"
+
+	"github.com/brahma-adshonor/gohook"
+	"github.com/fluid-cloudnative/fluid/pkg/utils/kubeclient"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 func TestAlluxioFileUtils_GetConf(t *testing.T) {
@@ -38,7 +56,7 @@ func TestAlluxioFileUtils_GetConf(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		stdout, err := AlluxioFileUtils{log: NullLogger{}}.GetConf(test.in)
+		stdout, err := AlluxioFileUtils{log: logf.NullLogger{}}.GetConf(test.in)
 		if stdout != test.out {
 			t.Errorf("input parameter is %s,expected %s, got %s", test.in, test.out, stdout)
 		}