+++ charts/jindofs/CHANGELOG.md
@@ -6,4 +6,21 @@ Support fuse critical pod
 0.8.0
 
 Change worker from Daemonset to Statefulset
+
+0.8.1
+
 Repair fuse MountPoint leak issue
+
+0.8.2
+
+Change podManagementPolicy from OrderedReady to Parallel
+
+0.8.3
+
+Add owner Reference
+
+0.8.4
+
+Support more Posix
+ 
+